requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
django-prometheus>=2.3.1
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
neo4j>=5.0.0
openai>=1.0.0
Pillow>=10.0.0
